Move test classes to the same package as the main classes
Having the test classes in the same package allows to test non-public methods directly.
